Create MergeSort.java
This is the sorting algorithm we are using for the STARS app console version. We are using this algorithm for testing purposes. The code in this file will not run as-is because it is part of another file. I separated out the sorting algorithm for academic purposes, and testing purposes.